Remove warnings

+++ maths/racine.h
@@ -27,7 +27,7 @@ namespace racine {
     template<typename Nombre>
     constexpr Nombre racine(Nombre n, size_t k) {
         static_assert(std::is_arithmetic<Nombre>::value, "Nombre doit être un type arithmetique.");
-        auto x = static_cast<Nombre>(std::pow(n, 1.0 / k));
+        auto x = static_cast<Nombre>(std::pow(n, 1.0 / static_cast<double>(k)));
         while (puissance::puissance<Nombre>(x, k) > n) --x;
         while (puissance::puissance<Nombre>(x + 1, k) <= n) ++x;
         return x;
